Widened the v1 value of Widened PacketEventMotionShort so that we can send easy messages such as "* Immune *" or "* Failed! *" appropriately instead of relying on truncation or wrap around.

+++ Sources/Dependencies/Shared/Packet/PacketEvent.h
@@ -103,7 +103,7 @@ namespace net {
 		PacketHeader header;
 		std::uint16_t object_id;
 		std::uint8_t dir;
-		std::uint8_t v1;
+		std::int16_t v1;
 		std::uint8_t v2;
 	};
